refactor(html): fold chunk-URL sentinel helpers into HtmlGenerator as static methods
`lib/html/htmlChunkUrl.js` was a single-file module exporting the
`makeHtmlChunkUrlSentinel` / `resolveHtmlChunkUrlSentinels` pair created
in the previous commit. Inline them onto `HtmlGenerator` as
`HtmlGenerator.makeChunkUrlSentinel` and `HtmlGenerator.resolveChunkUrlSentinels`
— that's the class responsible for emitting both sides of the
substitution (`_renderHtml` writes sentinels via dep templates and
resolves them on the JS-export path), so it's the natural home for the
sentinel format. `HtmlGenerator` doesn't import the HTML dependencies,
so the two dep templates (`HtmlInlineScriptDependency`,
`HtmlScriptSrcDependency`) can require it directly without introducing
a circular import. `HtmlModulesPlugin` already imports `HtmlGenerator`
and reaches the resolver through the same static method.

No behaviour change; just removes the standalone helper file.
