+++ lib/dependencies/HtmlInlineScriptDependency.js
@@ -4,7 +4,7 @@
 
 "use strict";
 
-const { makeHtmlChunkUrlSentinel } = require("../html/htmlChunkUrl");
+const HtmlGenerator = require("../html/HtmlGenerator");
 const makeSerializable = require("../util/makeSerializable");
 const ModuleDependency = require("./ModuleDependency");
 
@@ -110,7 +110,7 @@ HtmlInlineScriptDependency.Template = class HtmlInlineScriptDependencyTemplate e
 			// — done *before* the HTML's own `[contenthash]` is computed,
 			// so the HTML invalidates whenever the inline-script entry
 			// chunk's URL would change.
-			url = makeHtmlChunkUrlSentinel(chunk, "javascript");
+			url = HtmlGenerator.makeChunkUrlSentinel(chunk, "javascript");
 		}
 
 		// Insert ` src="…"` right after `<script` so the inline body is
