+++ lib/dependencies/HtmlScriptSrcDependency.js
@@ -9,7 +9,7 @@ const {
 	CSS_TYPE,
 	JAVASCRIPT_TYPE
 } = require("../ModuleSourceTypeConstants");
-const { makeHtmlChunkUrlSentinel } = require("../html/htmlChunkUrl");
+const HtmlGenerator = require("../html/HtmlGenerator");
 const makeSerializable = require("../util/makeSerializable");
 const ModuleDependency = require("./ModuleDependency");
 
@@ -387,7 +387,10 @@ HtmlScriptSrcDependency.Template = class HtmlScriptSrcDependencyTemplate extends
 		// `HtmlModulesPlugin#renderManifest` substitutes the sentinel
 		// before computing the HTML's own `[contenthash]`.
 		const entryContentHashType = isStylesheet ? "css" : "javascript";
-		const entryUrl = makeHtmlChunkUrlSentinel(entryChunk, entryContentHashType);
+		const entryUrl = HtmlGenerator.makeChunkUrlSentinel(
+			entryChunk,
+			entryContentHashType
+		);
 		source.replace(dep.range[0], dep.range[1] - 1, entryUrl);
 
 		if (dep.tagStart < 0 || dep.tagOpenEnd <= dep.tagStart) {
@@ -415,7 +418,7 @@ HtmlScriptSrcDependency.Template = class HtmlScriptSrcDependencyTemplate extends
 		 * @returns {string} a single sibling tag's HTML
 		 */
 		const buildSibling = (chunk, kind) => {
-			const url = makeHtmlChunkUrlSentinel(chunk, kind);
+			const url = HtmlGenerator.makeChunkUrlSentinel(chunk, kind);
 			if (kind === "css" && !isStylesheet) {
 				// Originating tag is `<script>` (or `<link rel=modulepreload>`)
 				// but this chunk is CSS — emit a fresh `<link>` rather than
