+++ lib/html/HtmlGenerator.js
@@ -14,10 +14,11 @@ const {
 } = require("../ModuleSourceTypeConstants");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const CssUrlDependency = require("../dependencies/CssUrlDependency");
-const { resolveHtmlChunkUrlSentinels } = require("./htmlChunkUrl");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("../../declarations/WebpackOptions").HtmlGeneratorOptions} HtmlGeneratorOptions */
+/** @typedef {import("../Chunk")} Chunk */
+/** @typedef {import("../Compilation")} Compilation */
 /** @typedef {import("../Compilation").DependencyConstructor} DependencyConstructor */
 /** @typedef {import("../CodeGenerationResults")} CodeGenerationResults */
 /** @typedef {import("../Dependency")} Dependency */
@@ -41,6 +42,126 @@ const { resolveHtmlChunkUrlSentinels } = require("./htmlChunkUrl");
 const JAVASCRIPT_AND_HTML_TYPES = new Set([JAVASCRIPT_TYPE, HTML_TYPE]);
 
 class HtmlGenerator extends Generator {
+	// Deferred-substitution sentinel for chunk URLs that need to land
+	// inside an HTML module's rendered output. Webpack runs dependency
+	// templates during `Compilation#codeGeneration()` — that's before
+	// `createHash()` populates `chunk.hash` and `chunk.contentHash[type]`,
+	// so any `[contenthash]` / `[chunkhash]` / `[fullhash]` in a chunk's
+	// filename template can't resolve at code-gen time. Instead, dep
+	// templates embed a chunk URL via `makeChunkUrlSentinel`, and
+	// `HtmlModulesPlugin#renderManifest` (which runs after chunk hashes
+	// are known) swaps every sentinel for
+	// `${PUBLIC_PATH_AUTO}<chunkFilename>` via `resolveChunkUrlSentinels`
+	// *before* hashing the HTML — so the HTML's own `[contenthash]`
+	// reflects the referenced chunks' hashes.
+	//
+	// Sentinel form:
+	// `__WEBPACK_HTML_CHUNK_URL__<hexChunkId>__<contentHashType>__END__`.
+	// `<hexChunkId>` is `chunk.id` UTF-8-encoded and hex-encoded so
+	// arbitrary id characters (including the `_` we use as a delimiter)
+	// can't conflict with the surrounding `__` separators.
+	// `<contentHashType>` is `javascript` or `css` — lowercase ASCII
+	// letters only, no encoding needed.
+
+	/**
+	 * Build a sentinel string that stands in for the chunk's emitted URL
+	 * during HTML code generation. Resolved by
+	 * `HtmlGenerator.resolveChunkUrlSentinels` once chunk hashes are
+	 * known.
+	 * @param {Chunk} chunk chunk whose URL should be embedded in the HTML
+	 * @param {"javascript" | "css"} contentHashType which `chunk.contentHash[type]` slice of the chunk the resolved URL should reference (`.js` vs `.css`)
+	 * @returns {string} sentinel string (or `data:,` if the chunk has no id yet, which should not happen in a well-formed compilation)
+	 */
+	static makeChunkUrlSentinel(chunk, contentHashType) {
+		const id = chunk.id;
+		if (id === null || id === undefined) return "data:,";
+		const hexId = Buffer.from(String(id), "utf8").toString("hex");
+		return `__WEBPACK_HTML_CHUNK_URL__${hexId}__${contentHashType}__END__`;
+	}
+
+	/**
+	 * Replace every sentinel produced by
+	 * `HtmlGenerator.makeChunkUrlSentinel` in `content` with
+	 * `${CssUrlDependency.PUBLIC_PATH_AUTO}<chunkFilename>`. Safe to call
+	 * from `HtmlModulesPlugin#renderManifest` (i.e. after `createHash()`
+	 * has populated every chunk's `chunk.hash` / `chunk.contentHash[type]`
+	 * and the compilation hash) — the chunk filename's `[contenthash]` /
+	 * `[chunkhash]` / `[fullhash]` placeholders all resolve there. Also
+	 * called from `HtmlGenerator#_renderHtml`'s JS-export path where
+	 * chunk hashes aren't computed yet; in that case `compilation.getPath`
+	 * may throw for templates carrying those placeholders, and the
+	 * sentinel is left in place so the later `renderManifest` pass can
+	 * still resolve it for the extracted HTML output. The
+	 * `PUBLIC_PATH_AUTO` prefix is left in place so the downstream
+	 * `undoPath` substitution in `HtmlModulesPlugin#renderManifest` still
+	 * applies.
+	 * @param {string} content HTML content possibly containing sentinels
+	 * @param {Compilation} compilation compilation
+	 * @returns {string} content with every sentinel replaced
+	 */
+	static resolveChunkUrlSentinels(content, compilation) {
+		if (!content.includes("__WEBPACK_HTML_CHUNK_URL__")) return content;
+		const outputOptions = compilation.outputOptions;
+		/** @type {Map<string, Chunk>} */
+		const chunksById = new Map();
+		for (const chunk of compilation.chunks) {
+			if (chunk.id !== null && chunk.id !== undefined) {
+				chunksById.set(String(chunk.id), chunk);
+			}
+		}
+		return content.replace(
+			/__WEBPACK_HTML_CHUNK_URL__([0-9a-f]+)__([a-z]+)__END__/g,
+			(match, hexId, contentHashType) => {
+				const chunkId = Buffer.from(hexId, "hex").toString("utf8");
+				const chunk = chunksById.get(chunkId);
+				if (!chunk) return "data:,";
+				let filenameTemplate;
+				if (contentHashType === "css") {
+					// CSS chunks use the CSS-specific filename template — same
+					// lookup `HtmlScriptSrcDependency` uses to emit
+					// `<link rel="stylesheet">`.
+					const CssModulesPlugin = require("../css/CssModulesPlugin");
+
+					filenameTemplate = CssModulesPlugin.getChunkFilenameTemplate(
+						chunk,
+						outputOptions
+					);
+				} else {
+					filenameTemplate =
+						chunk.filenameTemplate ||
+						(chunk.canBeInitial()
+							? outputOptions.filename
+							: outputOptions.chunkFilename);
+				}
+				try {
+					const filename = compilation.getPath(
+						/** @type {import("../TemplatedPathPlugin").TemplatePath} */
+						(filenameTemplate),
+						{
+							chunk,
+							contentHashType
+						}
+					);
+					return `${CssUrlDependency.PUBLIC_PATH_AUTO}${filename}`;
+				} catch (_) {
+					// `compilation.getPath` throws when the filename template
+					// contains a placeholder that isn't available in the
+					// current context — most commonly `[contenthash]` /
+					// `[chunkhash]` / `[fullhash]` called from `_renderHtml`'s
+					// JS-export path, where chunk hashes aren't computed yet.
+					// Leave the sentinel in place so the later
+					// `HtmlModulesPlugin#renderManifest` pass (run after
+					// `createHash()`) can still resolve it for the extracted
+					// HTML output. The JS export path will keep the sentinel
+					// string — JS callers of `require("./page.html")` who
+					// need dynamic chunk URLs should rely on `extract: true`
+					// instead.
+					return match;
+				}
+			}
+		);
+	}
+
 	/**
 	 * Creates an instance of HtmlGenerator.
 	 * @param {HtmlGeneratorOptions=} options generator options
@@ -284,7 +405,7 @@ class HtmlGenerator extends Generator {
 		// the HTML output path is the one that handles dynamic placeholders.
 		const compilation = generateContext.runtimeTemplate.compilation;
 		let content = /** @type {string} */ (source.source());
-		content = resolveHtmlChunkUrlSentinels(content, compilation);
+		content = HtmlGenerator.resolveChunkUrlSentinels(content, compilation);
 		content = content.split(CssUrlDependency.PUBLIC_PATH_AUTO).join(undoPath);
 		return content;
 	}
