+++ lib/html/HtmlModulesPlugin.js
@@ -326,7 +326,6 @@ class HtmlModulesPlugin {
 					makePathsRelative
 				} = require("../util/identifier");
 				const CssUrlDependency = require("../dependencies/CssUrlDependency");
-				const { resolveHtmlChunkUrlSentinels } = require("./htmlChunkUrl");
 
 				const autoPlaceholder = CssUrlDependency.PUBLIC_PATH_AUTO;
 
@@ -382,7 +381,7 @@ class HtmlModulesPlugin {
 							// placeholderContent) is what makes the HTML's
 							// own `[contenthash]` invalidate when the referenced
 							// chunks' filenames change.
-							const resolvedContent = resolveHtmlChunkUrlSentinels(
+							const resolvedContent = HtmlGenerator.resolveChunkUrlSentinels(
 								placeholderContent,
 								compilation
 							);
